Moving files to respective folders
