+++ src/modules/quiz/controllers/question.controller.ts
@@ -1,8 +1,8 @@
 import { Body, Controller, Post, UsePipes, ValidationPipe } from "@nestjs/common";
-import { CreateQuestionDto } from "./dto/create-question.dto";
-import { QuestionService } from "./question.service";
-import { Question } from "./question.entity";
-import { QuizService } from "./quiz.service";
+import { CreateQuestionDto } from "../dto/create-question.dto";
+import { QuestionService } from "../services/question.service";
+import { Question } from "../entities/question.entity";
+import { QuizService } from "../services/quiz.service";
 
 @Controller('question')
 export class QuestionController{
