+++ src/modules/quiz/quiz.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
-import { QuizController } from './quiz.controller';
-import { QuizService } from './quiz.service';
-import { QuizRepository } from './quiz.repository';
+import { QuizController } from './controllers/quiz.controller';
+import { QuizService } from './services/quiz.service';
+import { QuizRepository } from './repositories/quiz.repository';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Quiz } from './quiz.entity';
-import { QuestionController } from './question.controller';
-import { QuestionService } from './question.service';
-import { QuestionRepository } from './question.repository';
-import { Question } from './question.entity';
+import { Quiz } from './entities/quiz.entity';
+import { QuestionController } from './controllers/question.controller';
+import { QuestionService } from './services/question.service';
+import { QuestionRepository } from './repositories/question.repository';
+import { Question } from './entities/question.entity';
 
 @Module({
   controllers: [QuizController, QuestionController],
