+++ src/modules/quiz/services/quiz.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
-import { QuizRepository } from "./quiz.repository";
-import { CreateQuizDto } from "./dto/quiz.create.dto";
-import { Quiz } from "./quiz.entity";
+import { QuizRepository } from "../repositories/quiz.repository";
+import { CreateQuizDto } from "../dto/quiz.create.dto";
+import { Quiz } from "../entities/quiz.entity";
 
 @Injectable()
 export class QuizService {
